Fix circular dependency bug
