+++ src/server/models/Game.js
@@ -112,30 +112,6 @@ class Game {
 		}
 	}
 
-	static async getUserGames(id) {
-		try {
-			const rows = await db.any(sqlFile('game/get_user_games.sql'), { id: id });
-			return rows.map(row => {
-				for (let i = 1; i <= 4; i++) {
-					row[`player${i}`] = {
-						title: row[`player${i}Title`],
-						username: row[`player${i}Username`],
-						rating: row[`player${i}Rating`]
-					};
-					delete row[`player${i}Title`];
-					delete row[`player${i}Username`];
-					delete row[`player${i}Rating`];
-				}
-				return row;
-			});
-		} catch (err) {
-			if (!err.status) {
-				err.status = 500;
-			}
-			throw err;
-		}
-	}
-
 	static async getGameWithUsersByID(id) {
 		try {
 			const row = await db.oneOrNone(sqlFile('game/get_game_with_users_by_id.sql'), { id: id });
